+++ src/python/modernBERT/modernBERT_impute.py
@@ -9,7 +9,7 @@ from torch.utils.data import DataLoader
 from python.bed_io.bed import output_predictions
 from python.hmm.viterbi import build_pair_states, viterbi_decode
 from python.ps4g_io.torch_loaders import WindowIndexDatasetFromMatrix
-from modernbert import BERTImpute, BERTImputeConfig
+from python.modernBERT.modernBERT_model import BERTImputeConfig, BERTImpute
 
 
 def run_modernBERT_imputation(args, data, weights):
